requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
streamlit>=1.36.0
plotly>=5.22.0
